[refactor] change a constant
