+++ src/App.tsx
@@ -1,8 +1,8 @@
 import './App.css';
 import { Routes, Route } from 'react-router-dom'
-import { Intro } from './pages/Intro';
-import { Main } from './pages/Main';
-import { MainSelect } from './pages/MainSelect'
+import { Intro } from './pages/intro/Intro';
+import { Main } from './pages/main/Main';
+import { MainSelect } from './pages/main/MainSelect'
 import { Layout } from './routes/Layout';
 
 const App = () => {
@@ -10,12 +10,9 @@ const App = () => {
     <Routes>
       <Route element={<Layout />}>
         <Route path="/" element={<Intro></Intro>} />
-      </Route>
-      <Route element={<Layout />}>
         <Route path="/main" element={<Main></Main>} />
         <Route path="/main/select" element={<MainSelect></MainSelect>} />
       </Route>
-
     </Routes>
   );
 }
