+++ src/component/ThreeCard.tsx
@@ -22,7 +22,7 @@ export const ThreeCard = ({ isActive }: ThreeCardProps) => {
 
     return (
         <>
-            <div className={`absolute inset-0 bg-black bg-opacity-50 flex justify-center items-center ${isActive && visibleClass} `}>
+            <div className={`absolute inset-0 bg-black bg-opacity-50 flex justify-center items-center ${isActive && visibleClass}`}>
                 <div className='w-full h-4/5 flex flex-row justify-center relative'>
                     {cards.map((card, index) => (
                         <div key={index} className='flex-1 flex justify-center items-center relative'>
@@ -32,23 +32,20 @@ export const ThreeCard = ({ isActive }: ThreeCardProps) => {
                                 onClick={() => selectCard(card)}
                                 style={{
                                     transform: `perspective(1000px) rotateY(${rotationAngles[index].y}deg) rotateX(${rotationAngles[index].x}deg)`,
-                                    transformOrigin: 'center',
-                                    transition: 'all 0.1s',
                                     backgroundImage: `url(${card})`,
                                 }}>
                                 <div className='absolute inset-0 w-full h-full'
                                     style={{
                                         background: `radial-gradient(circle at ${overlayStyles[index].backgroundPosition}, rgba(255, 255, 255, 0.8) 0%, rgba(255, 255, 255, 0) 70%)`,
-                                        transition: 'all 0.1s',
                                         opacity: overlayStyles[index].opacity,
                                         filter: overlayStyles[index].filter,
                                     }}
                                 ></div>
-                            </div >
-                        </div >
+                            </div>
+                        </div>
                     ))}
-                </div >
-            </div >
+                </div>
+            </div>
 
             {selectedCard && selectedCardModal && <ThreeCardSelect isActive={isActive} card={selectedCard} close={() => {
                 setSelectedCardModal(false)
