+++ src/hooks/useCardSpread.ts
@@ -9,43 +9,35 @@ export const useCardSpread = (): IUseCardSpreadReturn => {
     const [visibleCards, setVisibleCards] = useState<boolean[]>(Array(12).fill(true));
 
     useEffect(() => {
-        setTimeout(() => {
-            setIsAnimated(true);
-        }, 500);
+        const timer = setTimeout(() => setIsAnimated(true), 500);
+        return () => clearTimeout(timer);
     }, []);
 
     useEffect(() => {
-        if (isAnimated) {
-            setPositions(spread());
-        } else {
+        const initialPositions = Array.from({ length: 12 }, (_, i) => ({ x: i, y: i }));
+        setPositions(isAnimated ? spread() : initialPositions);
+        if (!isAnimated) {
             setVisibleCards(Array(12).fill(true));
-            setPositions(Array.from({ length: 12 }, (_, i) => ({ x: i, y: i })));
             setSelectedCards([]);
         }
     }, [isAnimated]);
 
-    const spread = (): ISpreadCardPosition[] => {
-        return Array.from({ length: 12 }, (_, i) => {
-            const angle = (180 / 11 * i) * (Math.PI / 180);
-            return {
-                x: 600 * Math.cos(angle),
-                y: 200 * Math.sin(angle), //(x: 가로축, y 세로축)
-            };
-        });
-    };
+    const spread = (): ISpreadCardPosition[] => Array.from({ length: 12 }, (_, i) => {
+        const angle = (180 / 11 * i) * (Math.PI / 180);
+        return { x: 600 * Math.cos(angle), y: 200 * Math.sin(angle) };
+    });
 
     const handleCardClick = (index: number): void => {
-        if (selectedCards.length < 3) {
+        const MAX_SELECTED_CARD = 3
+        if (selectedCards.length < MAX_SELECTED_CARD) {
             setSelectedCards([...selectedCards, index]);
             setVisibleCards(visibleCards.map((visible, idx) => idx === index ? false : visible));
         }
     };
 
-    const resetAnimation = (): void => {
-        if (isAnimated) {
-            setIsAnimated(false);
-            setTimeout(() => setIsAnimated(true), 1500);
-        }
+    const resetAnimation = () => {
+        setIsAnimated(false);
+        setTimeout(() => setIsAnimated(true), 1500);
     };
 
     return { isAnimated, positions, selectedCards, visibleCards, handleCardClick, resetAnimation };
