+++ src/pages/main/Info.tsx
